pyproject.toml
[build-system]
requires = ["setuptools>=58", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-ant"
dynamic = ["version"]
description = "A lightweight iterative harness agent framework."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Nano Ant Contributors", email = "opensource@nano-ant.dev" }
]
urls = { Homepage = "https://github.com/example/nano-ant" }
dependencies = [
  "PyYAML>=6.0"
]
keywords = ["agent", "llm", "harness", "workflow", "automation"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
http = ["httpx>=0.27"]
dev = ["build>=1.2.2", "twine>=5.1.1"]

[project.scripts]
nano-ant = "nano_ant.cli:main"
ant = "nano_ant.cli:main"

[tool.setuptools.dynamic]
version = { attr = "nano_ant.__version__" }

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["nano_ant*"]

[tool.setuptools.package-data]
nano_ant = ["prompts/*.txt"]
